Combine buffers before transferring
