+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/playback/caching/DiskFileCacheDataSource.kt
@@ -10,8 +10,8 @@ import com.google.android.exoplayer2.upstream.TransferListener
 import com.lasthopesoftware.bluewater.client.browsing.items.media.files.cached.access.ICachedFilesProvider
 import com.lasthopesoftware.bluewater.client.browsing.items.media.files.cached.stream.CacheOutputStream
 import com.lasthopesoftware.bluewater.client.browsing.items.media.files.cached.stream.supplier.ICacheStreamSupplier
+import com.lasthopesoftware.bluewater.shared.drainQueue
 import com.lasthopesoftware.bluewater.shared.promises.NoopResponse.Companion.ignore
-import com.lasthopesoftware.bluewater.shared.promises.extensions.keepPromise
 import com.lasthopesoftware.bluewater.shared.promises.extensions.toPromise
 import com.lasthopesoftware.bluewater.shared.promises.extensions.unitResponse
 import com.lasthopesoftware.bluewater.shared.promises.toFuture
@@ -89,22 +89,23 @@ class DiskFileCacheDataSource(
 
 		private fun processQueue() : Promise<Unit> = synchronized(activePromiseSync) {
 			activePromise.eventually {
-				buffers.poll()
-					?.let { buffer ->
-						cachedOutputStream
-							.promiseTransfer(buffer)
-							.apply {
-								excuse {
-									logger.warn("An error occurred copying the buffer, closing the output stream", it)
-									clear()
-								}
-							}
-							.then {
-								processQueue().ignore()  // kick-off processing again, but don't wait for the result
-								Unit
-							}
+				val concatenatedBuffer = buffers
+					.drainQueue()
+					.fold(Buffer()) { sink, source -> sink.also(source::readAll) }
+
+				if (concatenatedBuffer.exhausted()) Unit.toPromise()
+				else cachedOutputStream
+					.promiseTransfer(concatenatedBuffer)
+					.apply {
+						excuse {
+							logger.warn("An error occurred copying the buffer, closing the output stream", it)
+							clear()
+						}
 					}
-					.keepPromise(Unit)
+					.then {
+						processQueue().ignore()  // kick-off processing again, but don't wait for the result
+						Unit
+					}
 			}.also { activePromise = it }
 		}
