+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/playback/engine/preparation/PreparedPlayableFileQueue.kt
@@ -5,6 +5,7 @@ import com.lasthopesoftware.bluewater.client.playback.file.PositionedPlayableFil
 import com.lasthopesoftware.bluewater.client.playback.file.preparation.PlayableFilePreparationSource
 import com.lasthopesoftware.bluewater.client.playback.file.preparation.PreparedPlayableFile
 import com.lasthopesoftware.bluewater.client.playback.file.preparation.queues.IPositionedFileQueue
+import com.lasthopesoftware.bluewater.shared.drainQueue
 import com.namehillsoftware.handoff.promises.Promise
 import org.joda.time.Duration
 import org.slf4j.LoggerFactory
@@ -19,20 +20,7 @@ class PreparedPlayableFileQueue(
 	private var positionedFileQueue: IPositionedFileQueue) : SupplyQueuedPreparedFiles, Closeable {
 
 	companion object {
-		private val logger = LoggerFactory.getLogger(PreparedPlayableFileQueue::class.java)
-
-		private fun <T> Queue<T>.drainQueue(): Iterable<T> {
-			val queue = this
-			return Iterable {
-				iterator {
-					do {
-						val next = queue.poll()
-						if (next != null)
-							yield(next)
-					} while (next != null)
-				}
-			}
-		}
+		private val logger by lazy { LoggerFactory.getLogger(PreparedPlayableFileQueue::class.java) }
 	}
 
 	private val queueUpdateLock = ReentrantReadWriteLock()
