+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/shared/KotlinExtensions.kt
@@ -1,0 +1,16 @@
+package com.lasthopesoftware.bluewater.shared
+
+import java.util.*
+
+fun <T> Queue<T>.drainQueue(): Iterable<T> {
+	val queue = this
+	return Iterable {
+		iterator {
+			do {
+				val next = queue.poll()
+				if (next != null)
+					yield(next)
+			} while (next != null)
+		}
+	}
+}
